WIP: Start 1.21.10 support

+++ ManagedServer/Entities/Types/Entity.cs
@@ -156,6 +156,7 @@ public class Entity : MappedTaggable, IViewable, IFeatureScope {
     protected virtual void Tick() {
         // Check for changed properties
         if (_lastVelocity != Velocity) {
+            // TODO: Velocity unit might have changed?
             SendToViewers(new ClientBoundSetEntityVelocityPacket {  // protocol velocity is in 8000ths of a block per tick (50ms)
                 EntityId = NetId,
                 Velocity = new Vec3<short>((short)(Velocity.X * 8000), (short)(Velocity.Y * 8000), (short)(Velocity.Z * 8000)),
@@ -174,7 +175,7 @@ public class Entity : MappedTaggable, IViewable, IFeatureScope {
             Yaw = Yaw,
             HeadYaw = HeadYaw,
             Data = 0,
-            Velocity = Vec3<short>.Zero
+            Velocity = Vec3<double>.Zero
         }, new ClientBoundSetEntityMetadataPacket {
             EntityId = NetId,
             Meta = Meta
